clean up recipe detail page and make previews
