+++ app/src/main/java/com/me/recipe/presentation/ui/recipe/RecipeScreen.kt
@@ -5,11 +5,7 @@ package com.me.recipe.presentation.ui.recipe
 import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionScope
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
@@ -17,18 +13,18 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.me.recipe.presentation.component.RecipeView
 import com.me.recipe.presentation.component.util.DefaultSnackbar
+import com.me.recipe.presentation.ui.recipe.components.RecipeDetail
 import com.me.recipe.util.compose.collectInLaunchedEffect
 import com.me.recipe.util.compose.use
 import kotlinx.coroutines.InternalCoroutinesApi
 import kotlinx.coroutines.launch
 
 @Composable
-fun RecipeScreen(
-    viewModel: RecipeViewModel = hiltViewModel(),
+internal fun RecipeScreen(
     sharedTransitionScope: SharedTransitionScope,
     animatedVisibilityScope: AnimatedVisibilityScope,
+    viewModel: RecipeViewModel = hiltViewModel(),
 ) {
     val (state, effect, event) = use(viewModel = viewModel)
     val snackbarHostState = remember { SnackbarHostState() }
@@ -51,18 +47,12 @@ fun RecipeScreen(
             }
         },
     ) { padding ->
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(padding)
-                .background(MaterialTheme.colorScheme.background),
-        ) {
-            RecipeView(
-                recipe = state.recipe,
-                isLoading = state.loading,
-                sharedTransitionScope = sharedTransitionScope,
-                animatedVisibilityScope = animatedVisibilityScope,
-            )
-        }
+        RecipeDetail(
+            recipe = state.recipe,
+            isLoading = state.loading,
+            sharedTransitionScope = sharedTransitionScope,
+            animatedVisibilityScope = animatedVisibilityScope,
+            modifier = Modifier.padding(padding),
+        )
     }
 }
