+++ gradle/libs.versions.toml
@@ -31,6 +31,7 @@ navigationTesting = "2.7.7"
 kotlinxCollectionsImmutable = "0.3.6"
 ksp= "1.9.23-1.0.19"
 spotless = "6.25.0"
+composeLintChecks = "1.3.1"
 
 
 [libraries]
@@ -71,6 +72,7 @@ datastore-preferences = { module = "androidx.datastore:datastore-preferences", v
 timber = { module = "com.jakewharton.timber:timber", version.ref = "timberVersion" }
 work = { module = "androidx.work:work-runtime-ktx", version.ref = "workKtx" }
 kotlinx-collections-immutable = { module = "org.jetbrains.kotlinx:kotlinx-collections-immutable", version.ref = "kotlinxCollectionsImmutable" }
+compose-lint-checks = { module = "com.slack.lint.compose:compose-lint-checks", version.ref = "composeLintChecks" }
